+++ src/components/playlist/components/playlist-button.js
@@ -5,12 +5,14 @@ const ClickableComponent = videojs.getComponent('ClickableComponent');
 
 // Create a common class for playlist buttons
 class PlaylistButton extends ClickableComponent {
+
   constructor(player, options) {
     // It is important to invoke the superclass before anything else,
     // to get all the features of components out of the box!
     super(player, options);
 
     const type = options.type;
+
     if (!type && type !== 'previous' && type !== 'next') {
       throw new Error('Type must be either \'previous\' or \'next\'');
     }