+++ src/components/recommendations-overlay/recommendations-overlay-item.js
@@ -3,9 +3,9 @@ import videojs from 'video.js';
 const ClickableComponent = videojs.getComponent('ClickableComponent');
 
 class RecommendationsOverlayItem extends ClickableComponent {
+
   setItem(item) {
     const { action, source } = item;
-
     this.source = source;
 
     const info = source.info();