+++ src/plugins/cloudinary/models/video-source.js
@@ -47,7 +47,8 @@ class VideoSource extends BaseSource {
       sourceTransformation,
       info,
       recommendations,
-      textTracks
+      textTracks,
+      withCredentials
     } = sliceAndUnsetProperties(
       options,
       'poster',
@@ -55,7 +56,8 @@ class VideoSource extends BaseSource {
       'sourceTransformation',
       'info',
       'recommendations',
-      'textTracks'
+      'textTracks',
+      'withCredentials'
     );
 
     super(publicId, options);
@@ -68,6 +70,7 @@ class VideoSource extends BaseSource {
     let _textTracks = null;
     this._type = 'VideoSource';
     this.isRawUrl = isRawUrl;
+    this.withCredentials = !!withCredentials;
 
     this.poster = (publicId, options = {}) => {
       if (!publicId) {
@@ -190,7 +193,7 @@ class VideoSource extends BaseSource {
       // if src is a url that already contains query params then replace '?' with '&'
       src += src.indexOf('?') > -1 ? queryString.replace('?', '&') : queryString;
 
-      return { type, src, cldSrc: this, isAdaptive: isAdaptive };
+      return { type, src, cldSrc: this, isAdaptive: isAdaptive, withCredentials: this.withCredentials };
     });
     if (isIe) {
       return srcs.filter(s => s.type !== 'video/mp4; codec="hev1.1.6.L93.B0"');
@@ -211,7 +214,7 @@ class VideoSource extends BaseSource {
       type = type ? `video/${type}` : null;
     }
 
-    return { type, src: url, cldSrc: this, isAdaptive };
+    return { type, src: url, cldSrc: this, isAdaptive, withCredentials: this.withCredentials };
   }
 }