+++ webpack/analyzer.config.js
@@ -1,11 +1,10 @@
 const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 const minified = require('./minified.config');
 
 
-module.exports = merge.smart(minified, {
+module.exports = merge(minified, {
   plugins: [
     new BundleAnalyzerPlugin()
   ]
 });
-