+++ webpack/dev.config.js
@@ -1,16 +1,15 @@
 const path = require('path');
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 const webpackCommon = require('./common.config');
 const env = require('../env');
 
 // Webpack plugins
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const DefinePlugin = require('webpack/lib/DefinePlugin');
-// const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 
-module.exports = merge.smart(webpackCommon, {
+module.exports = merge(webpackCommon, {
   devtool: 'inline-source-map',
   mode: 'development',